requirements.txt
flask
flask_restful
pandas
pandas_ta
ta
python-binance
matplotlib
